+++ src/main/java/com/example/modeladov1/model/Usuario.java
@@ -28,7 +28,7 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_usuario")
-    private Long id_usuario;
+    private int id_usuario;
 
     @Column(name="nombre")
     private String nombre;
@@ -49,11 +49,11 @@ public class Usuario {
 
     // Getters y setters
 
-    public Long getId_usuario() {
+    public int getId_usuario() {
         return id_usuario;
     }
 
-    public void setId_usuario(Long id_usuario) {
+    public void setId_usuario(int id_usuario) {
         this.id_usuario = id_usuario;
     }
